requirements.txt
tensorflow-gpu==2.4.1
tqdm<=4.59.0
